Update css language service: Fixes #27590 & 4/8 digit hex colors
