+++ extensions/css/client/src/colorDecorators.ts
@@ -4,7 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 'use strict';
 
-import { window, workspace, DecorationOptions, DecorationRenderOptions, Disposable, Range, TextDocument } from 'vscode';
+import * as parse from 'parse-color';
+import { window, workspace, DecorationOptions, DecorationRenderOptions, Disposable, Range, TextDocument, DocumentColorProvider, Color, ColorFormat, ColorInfo } from 'vscode';
 
 const MAX_DECORATORS = 500;
 
@@ -107,6 +108,12 @@ export function activateColorDecorations(decoratorProvider: (uri: string) => The
 				if (document && document.version === documentVersion && contentUri === document.uri.toString()) {
 					let decorations = ranges.slice(0, MAX_DECORATORS).map(range => {
 						let color = document.getText(range);
+						if (color[0] === '#' && (color.length === 5 || color.length === 9)) {
+							let c = Color.fromHex(color);
+							if (c) {
+								color = `rgba(${c.red}, ${c.green}, ${c.blue}, ${c.alpha})`;
+							}
+						}
 						return <DecorationOptions>{
 							range: range,
 							renderOptions: {
@@ -123,4 +130,54 @@ export function activateColorDecorations(decoratorProvider: (uri: string) => The
 	}
 
 	return Disposable.from(...disposables);
+}
+
+const CSSColorFormats = {
+	Hex: '#{red:X}{green:X}{blue:X}',
+	RGB: {
+		opaque: 'rgb({red}, {green}, {blue})',
+		transparent: 'rgba({red}, {green}, {blue}, {alpha:2f[0-1]})'
+	},
+	HSL: {
+		opaque: 'hsl({hue:d[0-360]}, {saturation:d[0-100]}%, {luminosity:d[0-100]}%)',
+		transparent: 'hsla({hue:d[0-360]}, {saturation:d[0-100]}%, {luminosity:d[0-100]}%, {alpha:2f[0-1]})'
+	}
+};
+
+function detectFormat(value: string): ColorFormat {
+	if (/^rgb/i.test(value)) {
+		return CSSColorFormats.RGB;
+	} else if (/^hsl/i.test(value)) {
+		return CSSColorFormats.HSL;
+	} else {
+		return CSSColorFormats.Hex;
+	}
+}
+
+export class ColorProvider implements DocumentColorProvider {
+
+	constructor(private decoratorProvider: (uri: string) => Thenable<Range[]>) { }
+
+	async provideDocumentColors(document: TextDocument): Promise<ColorInfo[]> {
+		const ranges = await this.decoratorProvider(document.uri.toString());
+		const result = [];
+		for (let range of ranges) {
+			let color;
+			const value = document.getText(range);
+			if (value[0] === '#') {
+				color = Color.fromHex(value);
+			} else {
+				const parsedColor = parse(value);
+				if (parsedColor && parsedColor.rgba) {
+					const [red, green, blue, alpha] = parsedColor.rgba;
+					color = new Color(red, green, blue, alpha);
+				}
+			}
+			if (color) {
+				const format = detectFormat(value);
+				result.push(new ColorInfo(range, color, format, [CSSColorFormats.Hex, CSSColorFormats.RGB, CSSColorFormats.HSL]));
+			}
+		}
+		return result;
+	}
 }
